requirements.txt
pytest==8.1.1
Django==5.0.6
djangorestframework==3.15.1
pytest-django==4.8.0